Add function instrumentation type accessors to function location type
Since the uprobe instrumentation is currently limited to function
entries, and since support for the instrumentation function return
is planned to be introduced at some point, it makes sense
to introduce an "instrumentation type" attribute on function
locations.

Currently, the only available instrumentation type is
"entry", which matches what is supported by the kernel tracer as of
2.11.

In the future, a RETURN and ENTRY_RETURN/BOTH instrumentation type
could be added without changing the default behavior of rules
such a userspace probe.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>