elf: support dynamic symbol table lookup
Background
==========
There may have two symbol tables in a shared object and binary. The
normal symbol table (.symtab) and the dynamic symbol table (.dynsym).

The normal symbol table contains lots of information, such static
linking data, but none of it is used at runtime. That why some
shared libraries are `stripped` to reduce the size of the file.
Stripping an object files removes the entire .symtab section of the elf
file, amongst other things.

The dynamic symbol table contains symbols that are needed for dynamic
linking of the shared object. The symbols in that section form a subset
of the symbols contained in the normal symbol section (before
stripping). The .dynsym section is left untouched when stripping a file
as it's needed at runtime.

Current limitation
==================
The current elf parsing implementation looks for the normal symbol
section (.symtab) to find the target symbol. If the .symtab is not
found, the parsing stops and returns that the symbol is not found. As
explained in the section above, a shared library might be stripped from
its normal symbol table but still have a dynamic symbol table (.dynsym)
containing the information of the target symbol. For example, on
distributions where the libc is stripped, the malloc symbol can only be
found in the .dynsym section

Solution
========
Look for the normal symbol section first, if it's found, used it to find
the symbol as usual. If the .symtab is absent, try to use the dynamic
symbol section instead.

This commit also adds a testcase for this feature.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>