Prevent channel buffer allocation larger than memory
Background
==========
Until recently (lttng-modules commit 1f0ab1e) it was possible to trigger
an Out-Of-Memory crash by creating a kernel channel buffer larger than
the currently usable memory on the system. The following commands was
triggering the issue on my laptop:
  lttng create
  lttng enable-channel -k --subbuf-size=100G --num-subbuf=1 chan0

The lttng-modules commit 1f0ab1e adds a verification based on an
estimate to prevent this from happening. Since this kernel tracer sanity
check is based on an estimate, it would safer to do a similar check on
the session daemon side.

Approach
========
Use the `/proc/meminfo` procfile to get an estimate of the current size
of available memory (using `MemAvailable`). The `MemAvailable` field was
added in the Linux kernel 3.14, so if it's absent we fallback to
verifying that the requested buffer is smaller than the physical memory
on the system.

Side effect
===========
This patch has the interesting side effect to alerting the user with an
error that buffer allocation has failed because of memory availability
in both --kernel and --userspace channel creation.

Drawback
========
The fallback check on older kernels is imperfect and is only to prevent
obvious user errors.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>