Fix: lttng_uri structure must be packed as it is used for IPC
The lttng_create_session commands send lttng_uri structures over
a UNIX socket. As such, the structure must be packed to preclude
the inclusion of any padding.

Moreover, the 'in_port_t' is replaced by uint16_t and PATH_MAX
is replaced by LTTNG_PATH_MAX to prevent conflicts if both ends
of the IPC are not build with the same toolchain/definitions.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>