lttng: clean-up the printout of snapshot outputs
The printout of snapshot outputs, when they are added or listed,
insistently uses "-1" and "0" to indicate that no "max-size"
parameter is specified.

This commit only prints the "max-size" settings when one is
actually specified and changes the spelling of "max-size" to
"max size".

Moreover, the unit (bytes) is printed when a max-size parameter
is specified on a snapshot output.

This changes the output of the listing of snapshot outputs as follows.

Before:
$ lttng snapshot list-output
Snapshot output list for session auto-20190327-133932
    [1] snapshot-1: /home/jgalar/lttng-traces/auto-20190327-133932 (max-size: 0)

After:
$ lttng snapshot list-output
Snapshot output list for session auto-20190327-133932
    [1] snapshot-1: /home/jgalar/lttng-traces/auto-20190327-133932

Creating a snapshot output without a max-size no longer indicates
"(max-size: -1)".

Before:
$ lttng snapshot add-output /tmp/keb
Snapshot output successfully added for session auto-20190327-134055
  [1] snapshot-1: /tmp/keb (max-size: -1)

After:
$ lttng snapshot add-output /tmp/keb
Snapshot output successfully added for session auto-20190327-134055
  [1] snapshot-1: /tmp/keb

Max sizes, when specified, are printed in bytes.

Before:
$ lttng snapshot add-output /tmp/keb --max-size=498394
Snapshot output successfully added for session auto-20190327-134055
  [1] snapshot-1: /tmp/keb (max-size: 498394)

After:
$ lttng snapshot add-output /tmp/keb --max-size=498394
Snapshot output successfully added for session auto-20190327-134055
  [1] snapshot-1: /tmp/keb (max size: 498394 bytes)

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>